+++ defmt/tests/ui/log-named-duplicate.rs
@@ -1,0 +1,3 @@
+fn main() {
+    defmt::info!("{x=u8}", x = 1, x = 2)
+}