+++ defmt/tests/ui/log-named-keyword.rs
@@ -1,0 +1,3 @@
+fn main() {
+    defmt::info!("{fn}")
+}