+++ defmt/tests/ui/log-named-positional-after-named.rs
@@ -1,0 +1,3 @@
+fn main() {
+    defmt::info!("{=u8} {x=u8}", x = 1, 5)
+}