+++ defmt/tests/ui/log-named-unused.rs
@@ -1,0 +1,3 @@
+fn main() {
+    defmt::info!("{x=u8}", x = 1, y = 2)
+}